backend/tcg/backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
python-jose[cryptography]
python-dotenv
passlib[bcrypt]